feat: support 8 decimals

+++ pkg/vault/contracts/lib/PoolConfigLib.sol
@@ -223,7 +223,7 @@ library PoolConfigLib {
             );
     }
 
-    function getTokenDecimalDiffs(PoolConfigBits config) internal pure returns (uint24) {
+    function getTokenDecimalDiffs(PoolConfigBits config) internal pure returns (uint40) {
         return
             uint24(
                 PoolConfigBits.unwrap(config).decodeUint(
@@ -254,7 +254,7 @@ library PoolConfigLib {
         return scalingFactors;
     }
 
-    function setTokenDecimalDiffs(PoolConfigBits config, uint24 value) internal pure returns (PoolConfigBits) {
+    function setTokenDecimalDiffs(PoolConfigBits config, uint40 value) internal pure returns (PoolConfigBits) {
         return
             PoolConfigBits.wrap(
                 PoolConfigBits.unwrap(config).insertUint(
@@ -288,8 +288,8 @@ library PoolConfigLib {
 
     // #endregion
 
-    // Convert from an array of decimal differences, to the encoded 24 bit value (only uses bottom 20 bits).
-    function toTokenDecimalDiffs(uint8[] memory tokenDecimalDiffs) internal pure returns (uint24) {
+    // Convert from an array of decimal differences, to the encoded 40-bit value (8 tokens * 5 bits/token).
+    function toTokenDecimalDiffs(uint8[] memory tokenDecimalDiffs) internal pure returns (uint40) {
         bytes32 value;
 
         for (uint256 i = 0; i < tokenDecimalDiffs.length; ++i) {
